refactor: remove validations test to separate file
